feat: refactor authentication and add code post module

+++ server/src/modules/user/service.ts
@@ -1,8 +1,9 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
+import { Like, Repository } from 'typeorm';
 import { UserDto } from './dto';
 import { User } from './entity';
+import { UserPayload } from './payload';
 
 @Injectable()
 export class UserService {
@@ -52,8 +53,17 @@ export class UserService {
     }
   }
 
-  async deleteUser(id: string): Promise<boolean> {
+  async deleteUser(id: string, userDto: UserDto): Promise<boolean> {
     try {
+      const user = await this.userRepository.findOne(id);
+
+      if (
+        user.id !== userDto.id ||
+        !userDto.roles.includes('ADMIN') ||
+        !userDto.roles.includes('MOD')
+      )
+        throw new Error(`You can't delete this account`);
+
       const result = await this.userRepository.delete({ id });
 
       return result.affected !== null;
@@ -62,17 +72,41 @@ export class UserService {
     }
   }
 
-  async updateUser(id: string, userDto: UserDto): Promise<UserDto> {
+  async updateUser(id: string, userPayload: UserPayload): Promise<UserDto> {
     try {
       const user = await this.userRepository.findOne(id);
 
       if (!user) return null;
 
-      Object.assign(user, userDto);
+      if (id !== user.id) throw new Error(`You can't modify this account`);
+
+      Object.assign(user, userPayload);
 
       const newUser = await this.userRepository.save(user);
 
       return new UserDto(newUser);
+    } catch (error) {
+      throw error;
+    }
+  }
+
+  async findUserByNameOrEmail(param: string): Promise<UserDto> {
+    try {
+      let user = null;
+
+      if (param.includes('@')) {
+        user = await this.userRepository.findOne({
+          email: Like(`%${param}%`),
+        });
+      } else {
+        user = await this.userRepository.findOne({
+          username: Like(`%${param}%`),
+        });
+      }
+
+      if (!user) return null;
+
+      return new UserDto(user);
     } catch (error) {
       throw error;
     }
